Improving and fixes: * Added depedencies at setup.py (Markdown and crcmod) * Replaced naming files from num by counting to generated crc hash by source * Removing prefixes of plugin name * Fix make png at PlantUML from tmp file * Added more examples of usage plugins * Changed checking <tag> with regexp because it did not work with jinja * Removed extract_format function, extracting format from regexp group * Improving speed with checking exist generated files

+++ mdx_ditaa.py
@@ -6,7 +6,8 @@ Based on mdx_graphviz.py, see it for details.
 """
 import markdown
 import markdown.preprocessors
-import os
+import crcmod.predefined
+import os, re
 import shutil
 import subprocess
 import tempfile
@@ -35,24 +36,24 @@ class DitaaPreprocessor(markdown.preprocessors.Preprocessor):
     """
 
     def __init__(self, ditaa):
-        self.ditaa = ditaa
         self.formatters = ["ditaa"]
+        self.ditaa    = ditaa
+        self.start_re = re.compile(r'^<(%s)>' % '|'.join(self.formatters))
+        self.end_re   = re.compile(r'^</(%s)>' % '|'.join(self.formatters))
+        self.crc64    = lambda x: crcmod.predefined.mkCrcFun('crc-64')(x)
 
     def run(self, lines):
-        start_tags = ["<%s>" % x for x in self.formatters]
-        end_tags = ["</%s>" % x for x in self.formatters]
-        graph_n = 0
         new_lines = []
         block = []
         in_block = None
         for line in lines:
-            line = line.rstrip('\r')
-            if line in start_tags:
+            start_tag = self.start_re.match(line)
+            end_tag   = self.end_re.match(line)
+            if start_tag:
                 assert(block == [])
-                in_block = self.extract_format(line)
-            elif line in end_tags:
-                new_lines.append(self.graph(graph_n, in_block, block))
-                graph_n = graph_n + 1
+                in_block = start_tag.group(1)
+            elif end_tag:
+                new_lines.append(self.graph(in_block, block))
                 block = []
                 in_block = None
             elif in_block in self.formatters:
@@ -62,31 +63,30 @@ class DitaaPreprocessor(markdown.preprocessors.Preprocessor):
         assert(block == [])
         return new_lines
 
-    def extract_format(self, tag):
-        fmt = tag[1:-1]
-        assert(fmt in self.formatters)
-        return fmt
-
-    def graph(self, n, kind, lines):
+    def graph(self, kind, lines):
         """Generates a graph from lines and returns a string containing n
         image link to created graph.
         """
+
+        code = "\n".join(lines)
+        name = self.crc64(code)
+
         assert(kind in self.formatters)
-        tmp = tempfile.NamedTemporaryFile()
-        tmp.write("\n".join(lines))
-        tmp.flush()
-        cmd = "%s %s %s" % (
-            os.path.join(self.ditaa.config["BINARY_PATH"], kind),
-            self.ditaa.config["ARGUMENTS"], tmp.name)
-        p = subprocess.Popen(cmd, shell=True, stdin=subprocess.PIPE,
-                             stdout=subprocess.PIPE, close_fds=True)
-        p.wait()
-        filepath = "%s%s%s.png" % (self.ditaa.config["WRITE_IMGS_DIR"],
-                                   kind, n)
-        shutil.copyfile(tmp.name + ".png", filepath)
-        output_path = "%s%s%s.png" % (self.ditaa.config["BASE_IMG_LINK_DIR"],
-                                      kind, n)
-        return "![Ditaa chart %s](%s)" % (n, output_path)
+        filepath = "%s%s.png" % (self.ditaa.config["WRITE_IMGS_DIR"], name)
+        if not os.path.exists(filepath):
+            tmp = tempfile.NamedTemporaryFile()
+            tmp.write(code)
+            tmp.flush()
+            cmd = "%s %s %s" % (
+                os.path.join(self.ditaa.config["BINARY_PATH"], kind),
+                self.ditaa.config["ARGUMENTS"], tmp.name)
+            p = subprocess.Popen(cmd, shell=True, stdin=subprocess.PIPE,
+                                 stdout=subprocess.PIPE, close_fds=True)
+            p.wait()
+            shutil.copyfile(tmp.name + ".png", filepath)
+
+        output_path = "%s%s.png" % (self.ditaa.config["BASE_IMG_LINK_DIR"], name)
+        return "![Ditaa chart %s](%s)" % (name, output_path)
 
 
 def makeExtension(configs=None):
